requirements.txt
praw
google-generativeai
python-telegram-bot
python-dotenv